requirements.txt
aiofiles==24.1.0
aiomysql==0.2.0
APScheduler==3.11.0
discord.py==2.4.0
httpx==0.27.2
